LFS-105: UI to browse NCI Thesaurus terms as a tree Refactor common request functions to util.jsx
